+++ CHANGELOG.rst
@@ -9,6 +9,14 @@ and this project adheres to `Semantic Versioning`_.
 .. _Keep a Changelog: https://keepachangelog.com/en/1.0.0/
 .. _Semantic Versioning: https://semver.org/spec/v2.0.0.html
 
+[Unreleased]
+------------
+
+Added
+^^^^^
+- Create train directory with sequential numbering.
+
+
 [0.1.14] - 2024-10-16
 ---------------------
 
@@ -42,7 +50,7 @@ Fixed
 - Mean Average Precision calculation fixed.
 
 Removed
--------
+^^^^^^^
 - Saving dictionary of results.
 - Confusion matrix for each batch size.