+++ CHANGELOG.rst
@@ -17,6 +17,9 @@ Added
 - Create train directory with sequential numbering.
 - Saving loss csv file for each epoch.
 - Improved verbose metrics display during training.
+- Resume training function.
+- Plot loss graphs from dataframe function.
+- Example with command line variables for training and resume training.
 
 
 [0.1.14] - 2024-10-16