+++ CHANGELOG.rst
@@ -15,12 +15,15 @@ and this project adheres to `Semantic Versioning`_.
 Added
 ^^^^^
 - Create train directory with sequential numbering.
-- Saving loss csv file for each epoch.
 - Improved verbose metrics display during training.
 - Resume training function.
-- Plot loss graphs from dataframe function.
 - Example with command line variables for training and resume training.
 
+Changed
+^^^^^^^
+- Saving loss csv file for each epoch in training model.
+- Training model with control flow of resume training.
+- Plot loss graphs from dataframe function.
 
 [0.1.14] - 2024-10-16
 ---------------------